feat: copy placed in its own file
